Extract equalProps shallow comparison util
