+++ src/client/visualizations/heat-map/heatmap-rectangles.tsx
@@ -22,7 +22,7 @@ import { ConcreteSeries } from "../../../common/models/series/concrete-series";
 import { SPLIT } from "../../config/constants";
 import { HeatMapRectangleRow } from "./heatmap-rectangle-row";
 import "./heatmap-rectangles.scss";
-import { equalProps } from "./utils/equal-props";
+import { equalProps } from "../../utils/equal-props/equal-props";
 import { ColorScale, LinearScale } from "./utils/scales";
 
 export interface HeatMapRectanglesProps {
